LayoutTests/media/broken-video.html

<video controls></video>
<p>Test that QuickTime file with broken content generates an error.<p>
<script src=video-test.js></script>
<script>
video.src = "content/garbage.mp4";
waitForEvent("error", function () {
    test("video.error");
    endTest();
});
</script>
