LayoutTests/media/broken-video-expected.txt

Test that QuickTime file with broken content generates an error.

EVENT(error)
TEST(video.error) OK
END OF TEST
